Migrated RestClientConfiguration tests.
